### 2024-03-18 d9db621
[pre-commit.ci] pre-commit autoupdate
updates:
- [github.com/psf/black: 23.12.1 → 24.3.0](https://github.com/psf/black/compare/23.12.1...24.3.0)
- [github.com/kynan/nbstripout: 0.6.1 → 0.7.1](https://github.com/kynan/nbstripout/compare/0.6.1...0.7.1)
- [github.com/nbQA-dev/nbQA: 1.7.1 → 1.8.4](https://github.com/nbQA-dev/nbQA/compare/1.7.1...1.8.4)
