+++ .pre-commit-config.yaml
@@ -5,15 +5,15 @@ repos:
   -   id: end-of-file-fixer
   -   id: trailing-whitespace
 - repo: https://github.com/psf/black
-  rev: 23.12.1
+  rev: 24.3.0
   hooks:
   -   id: black
 - repo: https://github.com/kynan/nbstripout
-  rev: 0.6.1
+  rev: 0.7.1
   hooks:
     - id: nbstripout
 - repo: https://github.com/nbQA-dev/nbQA
-  rev: 1.7.1
+  rev: 1.8.4
   hooks:
     - id: nbqa-black
     #- id: nbqa-isort
