### 2024-03-27 #241 @editorialbot
[REVIEW]: Learning Machine Learning with Lorenz-96
**Submitting author:** @dhruvbalwada (<a href="http://orcid.org/0000-0001-6632-0187">Dhruv Balwada</a>)
**Repository:** https://github.com/m2lines/L96_demo
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.3
