### 2024-05-06 m2lines/L96_demo#150 @AnonymousFool
Conda fails to solve environment - Windows
### Issue
When setting up the environment for building and executing the notebooks, I encountered a LibMambaUnsatisfiableError due to a conflict in environment.yaml.

### Context
I uninstalled anaconda and performed a fresh install. Through the anaconda prompt, I opened the folder containing the local repository and attempted to create a new environment from environment.yaml as described in the README. conda threw an error, deeming the packages nomkl and pytorch incompatible due to pytorch requiring mkl. The full output of the command can be found in a pastebin [here](https://pastebin.com/r2ep8byZ).

### Workaround
I created an empty environment and manually conda installed each package in environment.yaml except torch and torchvision, which I instead pip installed. I then successfully built and executed the notebooks.
