### 2024-05-07 m2lines/L96_demo#150 @IamShubhamGupto
@dhruvbalwada do you happen to have a windows machine laying around? Id like to work on that for some time.

Mac install logs -
[mac.log](https://github.com/m2lines/L96_demo/files/15237724/mac.log)

For whatever reason conda on linux and Mac are happy with the current `environment.yaml` but windows cannot seem to resolve it.

I have a really old windows laptop so it's just taking forever to get anything done. ill look into alternate ways of testing environments until then
