+++ environment.yaml
@@ -1,6 +1,6 @@
 # to turn this into a conda-linux-64.lock file, run:
 # conda-lock lock --mamba -f environment.yaml -p linux-64
-name: L96M2lines2
+name: L96M2lines
 channels:
   - conda-forge
   - defaults
