### 2024-05-07 m2lines/L96_demo#150 @IamShubhamGupto
@AnonymousFool could you confirm if you were successfully able to download and install `pysr` on windows ? my Mac and linux systems seems happy again but windows isn't. Might be a localized issue but wanted to check anyway
