### 2024-06-17 m2lines/L96_demo#155 @dhruvbalwada
Creating initial release
@IamShubhamGupto do you know how to solve the first remark that the reviewer made [here](https://github.com/openjournals/jose-reviews/issues/241#issuecomment-2067800371) about creating an initial release?
