+++ README.md
@@ -6,10 +6,9 @@
 
 This repository provides a set of notebooks to pedagogically introduce the reader to the problem of parameterization in the climate sciences and how machine learning may be used to address it.
 
-The original goal for these notebooks in this Jupyter book was for our [M2LInES](https://m2lines.github.io/) team to work together and learn from each other; in particular, to get up to speed on the key scientific aspects of our collaboration (parameterizations, machine learning, data assimilation, uncertainty quantification) and to develop new ideas. 
+The original goal for these notebooks in this Jupyter book was for our [M2LInES](https://m2lines.github.io/) team to work together and learn from each other; in particular, to get up to speed on the key scientific aspects of our collaboration (parameterizations, machine learning, data assimilation, uncertainty quantification) and to develop new ideas. Now this material is presented here for anyone to learn from. The primary audience for this guide is researchers and students trained in climate science wanting to be introduced to machine learning or trained in machine learning and want to get acquainted with the parameterization problem in climate sciences. Since the book addresses people from multiple fields the level of pre-requisites required is minimal; a basic understanding of Python and some experience with PDEs or dynamical systems and solving them numerically (an introductory course in numerical methods) can be helpful. This book could be used as a teaching tool, for self-study, or as a reference manual. 
 
-**Statement of need**: Now this material is presented here for anyone to learn from. The primary audience for this guide is researchers and students trained in climate science wanting to be introduced to machine learning or trained in machine learning and want to get acquainted with the parameterization problem in climate sciences. Since the book addresses people from multiple fields the level of pre-requisites required is minimal; a basic understanding of Python and some experience with PDEs or dynamical systems and solving them numerically (an introductory course in numerical methods) can be helpful. This book could be used as a teaching tool, for self-study, or as a reference manual. 
-
+The easiest way to read the content (non-interactively) is to view it through the book's [website](https://m2lines.github.io/L96_demo/intro.html). For more interactive experience either use the binder link provided above, or setup the appropriate environments on your own machine and interact with each notebook indivdually.
 
 ## Structure and Organization of the Repo
 
@@ -90,8 +89,15 @@ To remove cached notebooks, run
 $ jcache cache remove -p _build/.jupyter_cache
 ```
 
-## Contributing
+## Contributing and reporting problems
 
+If you find any problems or mistakes in the material, think something is not clear, or spot errors in the codes, please [open a new issue](https://github.com/m2lines/L96_demo/issues/new) to report these or seek help. 
+
+Also, we welcome any contributions that you would like to make. These can come in the form of:
+- Providing solutions to any errors or clarity issues you or others may have spotted and reported on the issues page. 
+- Suggest and create new notebooks for any additional concepts that are not currently covered.
+
+These contributions can also be made by [opening a new issue](https://github.com/m2lines/L96_demo/issues/new) and starting a discussion about what you would like to contribute, and eventually submitting changes in the form of a new pull request.
 
 ### Pre-commit
 
@@ -168,7 +174,7 @@ $ git fetch upstream
 $ git merge upstream/main
 ```
 
-Ideally you will not have any merge conflicts.
+Ideally, you will not have any merge conflicts.
 You are now ready to make a new feature branch.
 
 ## References
