### 2024-06-17 m2lines/L96_demo#155 @dhruvbalwada
In this process can we also delete the branches that have been merged (except the ones we are actively using to address reviewer comments), so that there is a fresh start.
