+++ references.bib
@@ -1,5 +1,12 @@
 % L96 Jupyter Book
 
+@article{christensen2022parametrization,
+  title={Parametrization in weather and climate models},
+  author={Christensen, Hannah and Zanna, Laure},
+  year={2022},
+  publisher={Oxford University Press}
+}
+
 @article{Lorenz1995,
   title = {Predictability: a problem partly solved},
   journal = {Seminar on Predictability},
