+++ paper.md
@@ -158,10 +158,10 @@ The material in this Jupyter book is presented over five sections. The first sec
 The book was created by and as part of M2LInES, an international collaboration supported by Schmidt Futures, to improve climate models with scientific machine learning. The original goal for these notebooks in this Jupyter book was for our team to work together and learn from each other; in particular, to get up to speed on the key scientific aspects of our collaboration (parameterizations, machine learning, data assimilation, uncertainty quantification) and to develop new ideas. This was done as a series of tutorials, each of which was led by a few team members and occurred with a frequency of roughly once every 2 weeks for about 6-7 months. This Jupyter book is a collection of the notebooks used during these tutorials, which have only slightly been edited for continuity and clarity. Ultimately, we are happy to share these resources with the scientific community to introduce our research ideas and foster the use of machine learning techniques for tackling climate science problems.
 
 # Statement of Need
-Parameterization of sub-grid processes is a major challenge in climate science. The exact details of this problem are often very context dependent (@christensen2022parametrization), ... 
+Parameterization of sub-grid processes is a major challenge in climate science. The exact details of this problem are often very context dependent (@christensen2022parametrization), ...
 
-As described above, these notebooks were originally created to introduce non-domain experts to ideas from the parameterization aspects of climate modeling and how machine learning could be used to potentially address these. Now they have been adapted to act as a pedagogical tool for self-learning, be used as a reference manual, or for teaching some modules in an introductory class on machine learning. 
-The book is organized in sections that are relatively independent; with the exception that the first section provides a general overview to the parameterization problem in climate models. 
+As described above, these notebooks were originally created to introduce non-domain experts to ideas from the parameterization aspects of climate modeling and how machine learning could be used to potentially address these. Now they have been adapted to act as a pedagogical tool for self-learning, be used as a reference manual, or for teaching some modules in an introductory class on machine learning.
+The book is organized in sections that are relatively independent; with the exception that the first section provides a general overview to the parameterization problem in climate models.
 Each notebook covers material that can be discussed in roughly an hour-long lecture, and sections can be mixed and matched or ordered as needed depending on the overall learning objectives.
 
 
