### 2024-06-21 m2lines/L96_demo#155 @IamShubhamGupto
Hey @dhruvbalwada, Ive merged the previous PR as I think it address the concerns of the review. For creating a release, Ive started a draft release [here](https://github.com/m2lines/L96_demo/releases). I hope youre able to see it. If not, let me know if you have any preference for the release title and description. Ive provided very basic entries for now.

You can either edit the draft or share the contents with me and I'll publish the release.
