### 2024-06-21 m2lines/L96_demo#155 @IamShubhamGupto
> I think the title and description seem good to me. Do we need to do anything else before we publish? Now that the README has also been ingested, maybe it makes sense to publish? If we make some changes to read me in response to the reviews, will they automatically appear in the release?

unfortunately, we would have to manually re release a newer version say `v1.1` to incorporate further changes by the reviewers. Ill create a release now and if further changes are required, we can always bump up the version number
