### 2024-07-29 #241 @dhruvbalwada
@magsol  - Are the to-do items in the above list meant to be clickable?
