### 2024-07-29 #241 @IamShubhamGupto
@magsol - regarding the third and last todo
`Archive the release on Zenodo/figshare/etc and post the DOI here` -  does GitHub count? as of now we have `v1.0` released and with the changes we will most likely make it `v1.1`. should we archive / delete `v1.0`?

`Make sure that the license listed for the archive is the same as the software license.`. Is the archive referred here same as the one in the third todo?
