### 2024-07-29 #241 @IamShubhamGupto
version number `v1.0`
https://github.com/m2lines/L96_demo/releases/tag/v1.0
