### 2024-08-05 8427d36
[pre-commit.ci] pre-commit autoupdate
updates:
- [github.com/psf/black: 24.4.2 → 24.8.0](https://github.com/psf/black/compare/24.4.2...24.8.0)
- [github.com/nbQA-dev/nbQA: 1.8.5 → 1.8.7](https://github.com/nbQA-dev/nbQA/compare/1.8.5...1.8.7)
