+++ .pre-commit-config.yaml
@@ -5,7 +5,7 @@ repos:
   -   id: end-of-file-fixer
   -   id: trailing-whitespace
 - repo: https://github.com/psf/black
-  rev: 24.4.2
+  rev: 24.8.0
   hooks:
   -   id: black
 - repo: https://github.com/kynan/nbstripout
@@ -13,7 +13,7 @@ repos:
   hooks:
     - id: nbstripout
 - repo: https://github.com/nbQA-dev/nbQA
-  rev: 1.8.5
+  rev: 1.8.7
   hooks:
     - id: nbqa-black
     #- id: nbqa-isort
