### 2024-08-15 #241 @magsol
@dhruvbalwada I've made my final pass and opened a ticket with some minor edits that are needed for the paper, including a missing reference. Once you have those fixed, you'll need to generate a new version and Zenodo archive. Once those are finished, I'll be able to complete review!
