### 2024-08-16 4c3b38d
Update paper.md
Solving #158
