### 2024-08-16 54025b6
Update paper.bib
Fixing #158
