+++ paper.bib
@@ -11,3 +11,10 @@
   url = {https://www.ecmwf.int/node/10829},
   author = {Lorenz, E.N.}
 }
+
+@article{christensen2022parametrization,
+  title={Parametrization in weather and climate models},
+  author={Christensen, Hannah and Zanna, Laure},
+  year={2022},
+  publisher={Oxford University Press}
+}
