### 2024-08-21 #241 @magsol
~~Huh, it looks like it's still having a hard time reading the DOIs but in the PDF they show up fine *shrug*~~

@dhruvbalwada Do you see the reference check above? It's finding the DOIs associated with your two references. Please add these DOIs to the bib file so they can be incorporated into your PDF's references, then go ahead and regenerate everything and we should be good to go.
