### 2024-08-21 9af2abf
add paper dois to paper.bib
