+++ paper.bib
@@ -10,6 +10,7 @@
   address = {Shinfield Park, Reading},
   url = {https://www.ecmwf.int/node/10829},
   author = {Lorenz, E.N.}
+  doi = {10.1017/cbo9780511617652.004}
 }
 
 @article{christensen2022parametrization,
@@ -17,4 +18,5 @@
   author={Christensen, Hannah and Zanna, Laure},
   year={2022},
   publisher={Oxford University Press}
+  doi = {10.1093/acrefore/9780190228620.013.826}
 }
