+++ paper.bib
@@ -9,7 +9,7 @@
   organization = {ECMWF},
   address = {Shinfield Park, Reading},
   url = {https://www.ecmwf.int/node/10829},
-  author = {Lorenz, E.N.}
+  author = {Lorenz, E.N.},
   doi = {10.1017/cbo9780511617652.004}
 }
 
@@ -17,6 +17,6 @@
   title={Parametrization in weather and climate models},
   author={Christensen, Hannah and Zanna, Laure},
   year={2022},
-  publisher={Oxford University Press}
+  publisher={Oxford University Press},
   doi = {10.1093/acrefore/9780190228620.013.826}
 }
