### 2024-08-21 #241 @magsol
Yep!

iPhone’d

On Wed, Aug 21, 2024 at 18:53 Dhruv Balwada ***@***.***>
wrote:

> @dhruvbalwada <https://github.com/dhruvbalwada> Do you see the reference
> check above? It's finding the DOIs associated with your two references.
> Please add these DOIs to the bib file so they can be incorporated into your
> PDF's references, then go ahead and regenerate everything and we should be
> good to go.
>
> Sounds good, when you say "go ahead and regenerate everything and we
> should be good to go.", do you mean generate a new version number a new
> zenodo archive?
>
