### 2024-08-21 #241 @dhruvbalwada
Is there any way to not go through this again? Generating a new zenodo archive is a massive pain because zenodo is not automatically picking up all the authors, and their affiliations and orcids (I spent an hour each time doing this by hand). Or should I not worry about correcting them in this zenodo archive version?
