### 2024-08-21 #241 @magsol
If the DOIs aren’t part of the zenodo archive, then nothing would change if
you regenerated it, so yeah you wouldn’t need to.

If that’s not the case, let me know and I’ll check with Lorena and see what
our options are.

iPhone’d

On Wed, Aug 21, 2024 at 18:58 Dhruv Balwada ***@***.***>
wrote:

> Is there any way to not go through this again? Generating a new zenodo
> archive is a massive pain because zenodo is not automatically picking up
> all the authors, and their affiliations and orcids (I spent an hour each
> time doing this by hand). Or should I not worry about correcting them in
> this zenodo archive version?
>
