### 2024-08-23 #241 @editorialbot
:wave: @openjournals/jose-eics, this paper is ready to be accepted and published.
