### 2024-08-28 #241 @magsol
@dhruvbalwada Nope, only @openjournals/jose-eics can make the final acceptance. We just have to sit tight :)
