### 2024-09-11 #241 @dhruvbalwada
@magsol - I noticed that the tag @openjournals/jose-eics doesn't seem to link to anything. Is there anything else we can do to help push things along?
