### 2024-09-11 #241 @magsol
@dhruvbalwada Like the checkboxes in the editor checklist above, that link requires certain permissions to view; suffice to say it does link to the list of EiCs for JOSE.

That said, we still just have to sit tight. I'm pinging the EiCs but otherwise they're the only ones who can finalize the acceptance. I know it's taking awhile and I deeply appreciate your patience; we're almost there.
