### 2024-09-27 #241 @IamShubhamGupto
@magsol does JOSE has a publication frequency? such as monthly releases? If so, with the new month almost here, I was hoping we could see some progress on this end.

Thanks
