### 2024-09-30 #241 @magsol
@IamShubhamGupto There is no publication frequency with JOSE, it happens instantly once all the checks have been made. The simple fact that is we are experiencing a massive monthslong backlog that started with the COVID pandemic and which we are still struggling to get out from under. I understand wanting to get this published as soon as possible, and it will, but for now we just have to wait. I know this isn't the answer you want, but I do appreciate your patience as we work through the backlog.
