### 2024-10-01 #241 @magsol
Partially. The backlog extends well beyond JOSE, as the EiCs have many competing responsibilities.
