### 2024-10-10 #241 @labarba
The title and author list on the Zenodo archive do not match the paper title and authors. Please check. (On Zenodo you can just update the metadata, there's no need for a new archive version. Note that you may not want Zenodo to do automatic updates of versions with each release, and that Zenodo pulls from committers into the author list.)
