### 2024-10-10 #241 @labarba
License mismatch: I note that the Zenodo archive shows Creative Commons Attribution 4.0 International, while the GitHub repository shows [MIT license](https://github.com/m2lines/L96_demo/blob/main/LICENSE). Licenses should match.

Relatedly, the footer of the JupyterBook just has a copyright notice, with no license listed. This could mislead readers that visit only the rendered book to think the materials are all-rights-reserved. Suggest modifying the footer to add license info.

And since this submission has both narrative content _and_ code, you may consider dual licensing: MIT for code, CC-BY for text and figures.
