### 2024-10-10 #241 @dhruvbalwada
After updating the footer and licensing, would you like us to do another release and update the zenodo with it?

The reason Zenodo is not matching is because there are some authors that did not directly contribute into github and also because most authors don't have their right affiliations on github. This means that we have to manually update the zenodo meta data. We are happy to do this, but would like to do this only after all other changes have been made. Unless you know of some way to copy the author metadata from an older zenodo version - the first few versions have the right authors?? However, since the review process keeps asking for new version after new version, we just gave up on updating zenodo meta data for so many authors and decided to wait till the end.
