### 2024-10-11 #241 @labarba
> After updating the footer and licensing, would you like us to do another release and update the zenodo with it?

The licensing in Zenodo is selected on metadata, so it would not require a new version; it is up to you if you want to update the archive after minor edits like the footer (which just changes a config file).
