### 2024-10-11 #241 @IamShubhamGupto
@labarba @dhruvbalwada

Let me know if this preview is good for the website:

Its replicated across all pages, if all is good, ill make a PR
