### 2024-10-11 #241 @labarba
Your could list the copyright owners as well as the year, and then the license. Like:

(c) Copyright 2024 Dhruv Balwada, Ryan Abernathey, Shantanu Acharya, et al. — License: MIT for code, CC-BY for text and figures.
