### 2024-10-11 #241 @labarba
Doubly licensed Jupyter notebooks are the norm nowadays, but are not handled gracefully by hosting services. You could leave the CC-BY license in Zenodo (I think it admits only one, but haven't checked in a while.) In the GitHub repo, you could add a License notice in the README.
