### 2024-10-11 #241 @IamShubhamGupto
Once merged, we will have to update the version one more time on GitHub if im not wrong
