### 2024-12-24 #241 @labarba
Hi @dhruvbalwada — thank you for your patience!

It doesn't look like this [comment](https://github.com/openjournals/jose-reviews/issues/241#issuecomment-2405812598) was addressed:

> The title and author list on the Zenodo archive do not match the paper title and authors. Please check. (On Zenodo you can just update the metadata, there's no need for a new archive version. Note that you may not want Zenodo to do automatic updates of versions with each release, and that Zenodo pulls from committers into the author list.)

Can you update the title of the Zenodo archive so it matches the paper? Please also double  check the author list: it looks like Johanna Goldman is listed on Zenodo but not the paper. The author list should match.
