### 2024-12-25 #241 @dhruvbalwada
@labarba - Now the latest version should have all these things fixed.
Apologies for this, I think I got confused since there are multiple versions on zenodo, which don't automatically get updated metadata from previous version. Everything is fixed in latest version now.
