### 2024-12-26 #241 @labarba
Congratulations @dhruvbalwada – your JOSE paper is published! 🚀

And thank you to our Editor: @magsol Reviewers: @Micky774, @AnonymousFool – ya'll make this possible! 🙏
